requirements.txt
click==8.1.3
discord.py==2.2.3
Pillow==9.5.0
psutil==5.9.5
pycryptodome==3.17
pydantic==2.6.1
pyobf2==1.2.0
pywin32==306
Requests==2.31.0
WMI==1.5.1
urllib3==2.0.2
attrs==23.1.0

types-psutil
types-requests